refactoring frag ref and type

+++ utils/storage/src/interner.rs
@@ -35,7 +35,7 @@ macro_rules! gen_span_constants {
     };
 
     (@recur ($acc:expr) $name:ident => $repr:literal, $($rest:tt)*) => {
-        pub const $name: Ident = Ident(FragSlice::new(FragSliceAddr::new($acc as u64, 0, $repr.len() as u16)));
+        pub const $name: Ident = Ident(FragSlice::new(FragSliceAddr::new($acc as u32, 0, $repr.len() as u16)));
         gen_span_constants!(@recur ($acc + $repr.len()) $($rest)*);
     };
